Add spellcheck example
One of the benefits of implementing system call support from scratch is
that we're able to have embedded zip filesystem support which trickles
into libraries such as stdio, without unportable symbolic interposition.
It's also be great if we could say open("gs://bucket/object", O_RDONLY)
for seamless GCS, similar to Java NIO, but abstracted by the C library.

+++ libc/bits/hamming.c
@@ -19,6 +19,9 @@
 ╚─────────────────────────────────────────────────────────────────────────────*/
 #include "libc/bits/bits.h"
 
-unsigned long(hamming)(unsigned long x, unsigned long y) {
+/**
+ * Counts number of different bits.
+ */
+unsigned long hamming(unsigned long x, unsigned long y) {
   return popcount(x ^ y);
 }
